+++ AlchemistCode/Extensions/StringExtensions.cs
@@ -14,7 +14,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "card_portraits", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find card image path: " + path);
+        MainFile.Logger.Error("Could not find card image path: " + path);
         return Path.Join(MainFile.ResPath, "images", "card_portraits", "card.png");
     }
 
@@ -32,7 +32,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "powers", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find power image path: " + path);
+        MainFile.Logger.Error("Could not find power image path: " + path);
         return Path.Join(MainFile.ResPath, "images", "powers", "power.png");
     }
 
@@ -41,7 +41,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "powers", "big", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find big power image path: " + path);
+        MainFile.Logger.Error("Could not find big power image path: " + path);
         return Path.Join(MainFile.ResPath, "images", "powers", "big", "power.png");
     }
 
@@ -50,7 +50,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "relics", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find relic image path: " + path);
+        MainFile.Logger.Error("Could not find relic image path: " + path);
         return Path.Join(MainFile.ResPath, "images", "relics", "relic.png");
     }
 
@@ -59,7 +59,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "relics", "big", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find big relic image path: " + path);
+        MainFile.Logger.Error("Could not find big relic image path: " + path);
         return Path.Join(MainFile.ResPath, "images", "relics", "big", "relic.png");
     }
 
@@ -68,7 +68,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "potions", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find potion image path: " + path);
+        MainFile.Logger.Error("Could not find potion image path: " + path);
         return path;
     }
 
@@ -77,7 +77,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "potions", "outlines", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find potion outline path: " + path);
+        MainFile.Logger.Error("Could not find potion outline path: " + path);
         return path;
     }
 
@@ -86,7 +86,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "badges", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find badge image path: " + path);
+        MainFile.Logger.Error("Could not find badge image path: " + path);
         return path;
     }
 
@@ -100,7 +100,7 @@ public static class StringExtensions
         path = Path.Join(MainFile.ResPath, "images", "enchantments", path);
         if (ResourceLoader.Exists(path)) return path;
 
-        MainFile.Logger.Info("Could not find enchantment image path: " + path);
+        MainFile.Logger.Error("Could not find enchantment image path: " + path);
         return path;
     }
 }